Commit final

+++ MVCBlog/models/post.php
@@ -1,22 +1,20 @@
 <?php
 require_once('models/connection.php');
-session_start();
 
 // Un post d'un blog
 class Post {
 
-  // Un post a trois attributs
-  // ils sont déclarés publiques pour éviter la lourdeur des accesseurs&mutateurs
-  //  (et pouvoir utiliser $post->author directement)
-  // C'est bien sûr à éviter dans un vrai projet
+  // Un post a quatre attributs
   private $author;
+  private $title;
   private $content;
   private $id; //éventuellement NULL si pas encore d’id attribué
 
 
   // Creation d'un post
-  public function __construct(string $author, string $content, $id=NULL) {
+  public function __construct(string $author, string $title, string $content, $id=NULL) {
     $this->set('author', $author);
+    $this->set('title', $title);
     $this->set('content', $content);
     $this->set('id', $id);
   }
@@ -34,6 +32,9 @@ class Post {
       case 'author':
         return $this->author;
         break;
+      case 'title':
+        return $this->title;
+        break;
       case 'content':
         return $this->content;
         break;
@@ -50,6 +51,9 @@ class Post {
     switch ($var_name) {
       case 'author':
         $this->author = pg_escape_string($value);
+        break;
+      case 'title':
+        $this->title = pg_escape_string($value);
         break;
       case 'content':
         $this->content = pg_escape_string($value);
@@ -70,7 +74,7 @@ class Post {
     // La requête ayant été préparée, on remplace :id avec la valeur de $id
     $req->bindParam(':id', $id);
     $req->execute();
-    $var_ret = $req->fetchAll(PDO::FETCH_CLASS | PDO::FETCH_PROPS_LATE, 'Post', array('author', 'content', 'id'));
+    $var_ret = $req->fetchAll(PDO::FETCH_CLASS | PDO::FETCH_PROPS_LATE, 'Post', array('author', 'title', 'content', 'id'));
     return $var_ret[0];
   }
 
@@ -81,7 +85,8 @@ class Post {
     }
     $condition_ok = FALSE;
     if (array_key_exists('last_writen_post', $_SESSION)) {
-      $condition_ok = $_SESSION['last_writen_post']->get('author') !== $this->get('author') || $_SESSION['last_writen_post']->get('content') !== $this->get('content');
+      $post = unserialize($_SESSION['last_writen_post']);
+      $condition_ok = $post->get('author') !== $this->get('author') || $post->get('content') !== $this->get('content') || $post->get('title') !== $this->get('title');
     }
     else {
         $condition_ok = TRUE;
@@ -91,12 +96,13 @@ class Post {
     if ($condition_ok) {
       $db = Db::getInstance();
 
-      $req = $db->prepare('INSERT INTO posts VALUES (:id, :author, :content)');
+      $req = $db->prepare('INSERT INTO posts VALUES (:id, :author, :title, :content)');
       $req->bindParam(':id', $this->id);
       $req->bindParam(':author', $this->author);
+      $req->bindParam(':title', $this->title);
       $req->bindParam(':content', $this->content);
       $success = $req->execute();
-      $_SESSION['last_writen_post'] = clone $this;
+      $_SESSION['last_writen_post'] = serialize($this);
     }
     else {
       $success = False;
@@ -107,14 +113,19 @@ class Post {
   // Enregistre le post dans la base, et renvoie vrai si l'écriture a réussi
   public static function delete($id): bool {
     $db = Db::getInstance();
-
     //On évite les injections SQL:
     $id = intval($id);
+
+    $post = Post::find($id);
+    if ($post->get('author') != $_SESSION['login']) {
+      return false;
+    }
 
     $req = $db->prepare('delete FROM posts WHERE id=:id');
     $req->bindParam(':id', $id);
     // On transforme notre classe en tableau et on le passe en paramètre à execute
     $success = $req->execute();
+
     return $success;
   }
